test: refactor multimodal test
